Move ApiConfig to http
